The user input should be in a format of hours: minutes: and then seconds. You will start listening to the song as you will reach the time that has been set. To test your code set the time 2 or 3 minutes later from the time you are giving the user input.
Summary
This idea can be implemented in software applications also, so you now have an idea of what can be a good Python project other than just designing the User interface of an application.

I hope you liked this article on how to write a program to create an alarm with Python. Feel free to ask your valuable questions in the comments section below.